Do not store hashNext on disk
Instead of trusting the hashNest stored in the serialized CDiskBlockIndex
entries, use just the hashBestChain pointer to represent the active chain.

This means the block connection logic does not need to touch the block
index entries anymore, and will allow for easier separation of this database
in two independent parts (data about the block tree, and data about the
active chain).