Swap unordered_map for map in PartiallyDownloadedBlock::InitData
This saves an estimated 5% to 7% in the time-critical
mempool-scanning step. If there were fewer memory reads for SipHash
this could save a bunch more.

Note that we cant just blindly swap the map for an unordered_map
since we're using entirely untrusted indices, which could result
in a worst-case DoS issue...instead we check that each bucket does
not overfill the expected maximum bucket size, falling back to
regular block download if it one does.