Replace CombineSignatures with ProduceSignature
Instead of using CombineSignatures to create the final scriptSig or
scriptWitness of an input, use ProduceSignature itself.

To allow for ProduceSignature to place signatures, pubkeys, and scripts
that it does not know about, we use a new SigningProvider and
SignatureCreator which wrap a SignatureData (which contains sigs, pubkeys,
and scripts) to retrieve those unknown elements. The SigningProvider
also contains another SigningProvider which is used to create the
new signatures and get new scripts. The SignatureData gets its scripts,
pubkeys, and signatures from the data already in the scriptSig or from
the scriptSigs of transactions being merged.